CacheReader: Abstract CubeIterator interface
Surprisingly N=14 is not possible with 16GiB of memory
because at certain point of progress OS begins to swap
*something* and the process grinds to halt.
This happens even if *there is free-memory available* so something is
going haywire.

I found out that the culprit maybe that large (+3 GiB)
CacheReader memory mappings are being swapped out from the memory.
The OS is trying keeping the previously accessed memory in system memory
to our detriment.
For -t K threads we only need to have K Cubes from the cache-file
in memory at once.

Only way out of this problem is to not memory map the entire cache file
at once and instead read it Cube-by-Cube.
I think @nsch0e would have wanted to implement reading this way from
the beginning but he was missing the `mapped::file::readAt()`
that works with absolute file offsets and can read the file in parallel.

Currently FlatCache and CacheReader use the same CubeIterator
and ShapeRange types.
This is an problem for implementing better CubeIterator that reads
the Cubes one-by-one from a file because any changes to
these would break FlatCache that doesn't use cache files.

Start by adding abstract interfaces for CubeIterator and ShapeRange.

- ICubeIterator base class interface for Cube iterators
- CubeIterator the current implementation for ICubeIterator.
- CacheIterator type-erased proxy.
  This is needed to avoid disrupting the CubeIterator class users
  too much and make the type-erased iterator work in practice.
- IShapeRange base class interface.
- Make ICache::getCubesByShape() return reference to the IShapeRange.
- Adapt CubeIterator users to use CacheIterator instead.

Signed-off-by: JATothrim <jarmo.tiitto@gmail.com>